refactor(search): share channel name resolution with call repo
Lift batch_resolve_channel_names and the helper resolve_channel_name
into comms_db_client so both the call repo and the search enrichment
use the same logic. Search now resolves channel names for DM and
unnamed-private channels (previously the simple LEFT JOIN returned
null).
