--- rust/cloud-storage/.sqlx/query-341d86f1e9434f4beff9084ee0e9574f6d97248022183cf8680f7553f2f9ce81.json
@@ -1,65 +1,0 @@
-{
-  "db_name": "PostgreSQL",
-  "query": "\n        SELECT\n            cr.id AS \"call_id!\",\n            cr.created_by AS \"created_by!\",\n            cr.started_at AS \"started_at!\",\n            cr.ended_at AS \"ended_at!\",\n            cr.duration_ms AS \"duration_ms!\",\n            cc.name AS \"channel_name?\",\n            cr.custom_name AS \"custom_name?\",\n            EXISTS (\n                SELECT 1 FROM call_record_participants crp\n                WHERE crp.call_record_id = cr.id AND crp.user_id = $2\n            ) AS \"attended!\"\n        FROM call_records cr\n        LEFT JOIN comms_channels cc ON cc.id = cr.channel_id\n        WHERE cr.id = ANY($1)\n        ",
-  "describe": {
-    "columns": [
-      {
-        "ordinal": 0,
-        "name": "call_id!",
-        "type_info": "Uuid"
-      },
-      {
-        "ordinal": 1,
-        "name": "created_by!",
-        "type_info": "Text"
-      },
-      {
-        "ordinal": 2,
-        "name": "started_at!",
-        "type_info": "Timestamptz"
-      },
-      {
-        "ordinal": 3,
-        "name": "ended_at!",
-        "type_info": "Timestamptz"
-      },
-      {
-        "ordinal": 4,
-        "name": "duration_ms!",
-        "type_info": "Int8"
-      },
-      {
-        "ordinal": 5,
-        "name": "channel_name?",
-        "type_info": "Varchar"
-      },
-      {
-        "ordinal": 6,
-        "name": "custom_name?",
-        "type_info": "Text"
-      },
-      {
-        "ordinal": 7,
-        "name": "attended!",
-        "type_info": "Bool"
-      }
-    ],
-    "parameters": {
-      "Left": [
-        "UuidArray",
-        "Text"
-      ]
-    },
-    "nullable": [
-      false,
-      false,
-      false,
-      false,
-      false,
-      true,
-      true,
-      null
-    ]
-  },
-  "hash": "341d86f1e9434f4beff9084ee0e9574f6d97248022183cf8680f7553f2f9ce81"
-}
