+++ rust/cloud-storage/.sqlx/query-767162a3de69fd6df368e97dbf8ffdca791dfc0f1557adc7203b89ee8082251d.json
@@ -1,6 +1,6 @@
 {
   "db_name": "PostgreSQL",
-  "query": "\n            SELECT id, name, channel_type::text as \"channel_type!\"\n            FROM comms_channels\n            WHERE id = ANY($1)\n            ",
+  "query": "\n        SELECT id, name, channel_type::text as \"channel_type!\"\n        FROM comms_channels\n        WHERE id = ANY($1)\n        ",
   "describe": {
     "columns": [
       {
@@ -30,5 +30,5 @@
       null
     ]
   },
-  "hash": "6ca02c6511b3279813710cb93a275d215c3edeb60072739f71e5e7324772349c"
+  "hash": "767162a3de69fd6df368e97dbf8ffdca791dfc0f1557adc7203b89ee8082251d"
 }
