+++ rust/cloud-storage/.sqlx/query-a42ed7aaed8d1c0d84fe66d9ed2dc61ff75417b2134173c46cd1cdc171b8f726.json
@@ -1,6 +1,6 @@
 {
   "db_name": "PostgreSQL",
-  "query": "\n                SELECT channel_id, user_id\n                FROM comms_channel_participants\n                WHERE channel_id = ANY($1) AND left_at IS NULL\n                ",
+  "query": "\n            SELECT channel_id, user_id\n            FROM comms_channel_participants\n            WHERE channel_id = ANY($1) AND left_at IS NULL\n            ",
   "describe": {
     "columns": [
       {
@@ -24,5 +24,5 @@
       false
     ]
   },
-  "hash": "6dae1be229aa02fd589a494c2219f9ffafb95d71284c9062109b4fc7ccf80206"
+  "hash": "a42ed7aaed8d1c0d84fe66d9ed2dc61ff75417b2134173c46cd1cdc171b8f726"
 }
