+++ rust/cloud-storage/call/src/domain/mod.rs
@@ -1,6 +1,3 @@
-/// Channel name resolution logic.
-pub mod channel_name;
-
 /// Domain models for calls.
 pub mod models;
