+++ rust/cloud-storage/call/src/outbound/pg_call_repo.rs
@@ -8,6 +8,9 @@ mod test;
 use std::collections::{HashMap, HashSet};
 
 use chrono::Utc;
+use comms_db_client::channels::resolve_names::{
+    NameLookup, batch_resolve_channel_names, display_name, resolve_channel_name,
+};
 use entity_access::domain::models::AccessLevel;
 use filter_ast::Expr;
 use item_filters::ast::{LiteralTree, call::CallLiteral};
@@ -17,7 +20,6 @@ use models_permissions::share_permission::channel_share_permission::ChannelShare
 use sqlx::PgPool;
 use uuid::Uuid;
 
-use crate::domain::channel_name::{NameLookup, display_name, resolve_channel_name};
 use crate::domain::models::{
     AddParticipantError, Call, CallParticipant, CallRecord, CallRecordParticipant,
     CallRecordPreview, CallRecordPreviewData, CallRecordTranscriptSegment, CustomSpeakerAssignment,
@@ -93,112 +95,6 @@ impl PgCallRepo {
     /// Create a new repo with the given connection pool.
     pub fn new(pool: PgPool) -> Self {
         Self { pool }
-    }
-
-    /// Batch-resolve display names for a list of channel ids from the
-    /// perspective of `viewer_user_id`. The viewer is only used to pick the
-    /// right "other person" for DM channels; it is not an authorization check.
-    ///
-    /// Channels the query can't find simply have no entry in the returned map.
-    async fn batch_resolve_channel_names<'a>(
-        &self,
-        channel_ids: &[Uuid],
-        viewer_user_id: MacroUserIdStr<'a>,
-    ) -> Result<HashMap<Uuid, String>, sqlx::Error> {
-        if channel_ids.is_empty() {
-            return Ok(HashMap::new());
-        }
-
-        let channel_info_rows = sqlx::query!(
-            r#"
-            SELECT id, name, channel_type::text as "channel_type!"
-            FROM comms_channels
-            WHERE id = ANY($1)
-            "#,
-            channel_ids,
-        )
-        .fetch_all(&self.pool)
-        .await?;
-
-        let channel_map: HashMap<Uuid, (Option<String>, String)> = channel_info_rows
-            .into_iter()
-            .map(|r| (r.id, (r.name, r.channel_type)))
-            .collect();
-
-        // Channels that need participant-derived names (DMs, unnamed privates).
-        let needs_participants: Vec<Uuid> = channel_map
-            .iter()
-            .filter(|(_, (name, ct))| {
-                ct == "direct_message"
-                    || (ct == "private" && name.as_ref().is_none_or(|n| n.trim().is_empty()))
-            })
-            .map(|(id, _)| *id)
-            .collect();
-
-        let (participant_map, name_lookup) = if needs_participants.is_empty() {
-            (HashMap::new(), NameLookup::new())
-        } else {
-            let participant_rows = sqlx::query!(
-                r#"
-                SELECT channel_id, user_id
-                FROM comms_channel_participants
-                WHERE channel_id = ANY($1) AND left_at IS NULL
-                "#,
-                &needs_participants
-            )
-            .fetch_all(&self.pool)
-            .await?;
-
-            let mut part_map: HashMap<Uuid, Vec<MacroUserIdStr<'static>>> = HashMap::new();
-            let mut all_user_ids = HashSet::new();
-            for row in participant_rows {
-                let uid = MacroUserIdStr::parse_from_str(&row.user_id)
-                    .expect("valid user id from db")
-                    .into_owned();
-                all_user_ids.insert(row.user_id);
-                part_map.entry(row.channel_id).or_default().push(uid);
-            }
-
-            let user_id_strings: Vec<String> = all_user_ids.into_iter().collect();
-            let name_rows = sqlx::query!(
-                r#"
-                SELECT u.id as user_profile_id, mui.first_name, mui.last_name
-                FROM macro_user_info mui
-                JOIN "User" u ON mui.macro_user_id = u.macro_user_id
-                WHERE u.id = ANY($1)
-                "#,
-                &user_id_strings
-            )
-            .fetch_all(&self.pool)
-            .await?;
-
-            let lookup: NameLookup = name_rows
-                .into_iter()
-                .filter_map(|row| {
-                    let name = display_name(row.first_name.as_deref(), row.last_name.as_deref())?;
-                    Some((row.user_profile_id, name))
-                })
-                .collect();
-
-            (part_map, lookup)
-        };
-
-        let mut resolved: HashMap<Uuid, String> = HashMap::with_capacity(channel_map.len());
-        for (channel_id, (name, channel_type)) in &channel_map {
-            let empty = Vec::new();
-            let participants = participant_map.get(channel_id).unwrap_or(&empty);
-            let resolved_name = resolve_channel_name(
-                channel_type,
-                name.as_deref(),
-                participants,
-                channel_id,
-                viewer_user_id.copied(),
-                &name_lookup,
-            );
-            resolved.insert(*channel_id, resolved_name);
-        }
-
-        Ok(resolved)
     }
 }
 
@@ -994,9 +890,8 @@ impl CallRepository for PgCallRepo {
                 .collect()
         };
 
-        let channel_names = self
-            .batch_resolve_channel_names(&unique_channel_ids, user_id)
-            .await?;
+        let channel_names =
+            batch_resolve_channel_names(&self.pool, &unique_channel_ids, user_id).await?;
 
         let previews = unique_call_ids
             .into_iter()
@@ -1181,9 +1076,8 @@ impl CallRepository for PgCallRepo {
                 .collect()
         };
 
-        let channel_names = self
-            .batch_resolve_channel_names(&unique_channel_ids, user_id.copied())
-            .await?;
+        let channel_names =
+            batch_resolve_channel_names(&self.pool, &unique_channel_ids, user_id.copied()).await?;
 
         for record in &mut records {
             record.channel_name = channel_names.get(&record.channel_id).cloned();
