+++ rust/cloud-storage/comms_db_client/src/channels/mod.rs
@@ -7,5 +7,6 @@ pub mod get_dm;
 pub mod get_mentions_for_channel;
 pub mod get_private;
 pub mod patch_channel;
+pub mod resolve_names;
 pub mod seed_channel;
 pub mod updated_at;
